refactor: file rename and folder restructure
